requirements.txt
anyascii==0.3.2
asgiref==3.8.1
beautifulsoup4==4.12.3
certifi==2024.12.14
charset-normalizer==3.4.1
defusedxml==0.7.1
Django==4.2.3
django-admin-interface==0.19.1
django-asset-definitions==1.0.0
django-ckeditor==6.7.2
django-colorfield==0.12.0
django-cors-headers==4.6.0
django-crispy-forms==1.14.0
django-dashboards==0.1.7
django-environ==0.11.2
django-filter==24.3
django-flat-responsive==2.0
django-flat-theme==1.1.4
django-js-asset==3.0.1
django-modelcluster==6.4
django-permissionedforms==0.1
django-taggit==6.1.0
django-treebeard==4.7.1
djangorestframework==3.15.2
draftjs-exporter==5.0.0
et_xmlfile==2.0.0
filetype==1.2.0
idna==3.10
l18n==2021.3
laces==0.1.1
mysqlclient==2.2.7
numpy==2.2.1
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
patsy==1.0.1
pillow==11.1.0
pillow_heif==0.21.0
plotly==5.24.1
plotly-express==0.4.1
python-dateutil==2.9.0.post0
pytz==2024.2
requests==2.32.3
scipy==1.15.0
six==1.17.0
soupsieve==2.6
sqlparse==0.5.3
statsmodels==0.14.4
sweetify==2.3.1
telepath==0.3.1
tenacity==9.0.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
wagtail==6.3.2
whitenoise==6.8.2
Willow==1.9.0
